+++ projects/ngx-single-select/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## 8.0.1
+- Added Keyboard Navigation
+- Added Search for listed items
+
 ## 8.0.0
 - initial release